tcfl.report.file_c: use Jinja2 for text, junit, etc
This rips off all the hardcoded text and junit_xml support and
replaces it with Jinja2 templates.

Jinja2 templates allows to create, with the same code path, multiple
formats (templates for text and junit.xml are provided) with much more
flexibiltiy and less horrible hacks.

The file_c becomes a simple walk over all defined templates, defining
keywords to be exported to Jinja2 for each template.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>