tcfl/zephyr-sanity: reduce timing chances of false negative
On highly loaded systems, there was a chance for a false negative to
creep in if the current task got delayed in between calling the three
target.expects() needed to verify a Zephyr sanity check testcase.

Thus, it would look like a timeout had happened, when the console
output clearly showed the three markers indicating a succesfult
testcase execution.

To reduce the chances of said problem happening, replace with setting
three expectations that will be checked almost at the same time.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>