ttbl.test_target: verify IP addresses belong to interconect's network
When there is access to the interconnect (it is in the same host),
verify the IP addresses given by a target fit in that interconnect's
network definition or error out.

This allows you not to spend two hours debugging why DHCP is not
assigning the pre-defined IP address like I just did.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>