ttbd/POS: script to setup distribution images
This script will take as input live filesystems or other images and
generate a directory tree with the contents of it as it would look in
an installed Linux distribution.

Eg: given the Fedora Live 29 iso, it will create a directory tree that
would be the contents of said Live iso when the system is executed.

The goal is to be able to rsync these directory trees into a root
filesystem as if they had been installed right off the ISO.

A few tweaks are done to simplify the testing environment.

Currently supported:

- Fedora
- Ubuntu
- Clear
- Yocto
- tcf-live
- debian?

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>