tcfl/POS: add client side of Provioning OS
tcfl.pos is a library with multiple helper functions that can be used
to provision a Linux target using a Provisioning OS.

The POS boots in the target and the script has access to all the
drives to install anything as it sees fit.

The general idea is to have any testcase start in the deployment phase
with ensuring the right image is installed in the target (eg: Fedora
29) as if it was freshly installed and then doing any additional setup
after booting to F29 (eg: installing and configuring apache) and then
runing the actual testcase.

The scripts in this library provide support for doing this and a way
to optimize it by using preinstalled distro images available in an
rsync server which can be flashed at quite high speed (a 1.1GB image
will flash from empty in < 1min on a 1Gbps link; if the image has been
barely modified, the image will be reset to new in less than a few
seconds).

Examples and documentation in further commits.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>