tcfl/zephyr: fix data harvester access to the console
Data harvester had stopped working properly because the file
descriptor it was getting had its file pointer at the end of the file.

This uncovered that it was being done wrong; then the harvester gets
the descriptor it shan't assume it is pointing to the beginning of the
file where the console has been captured.

So dup it and read the whole thing from the start. Document the caveat
in the expecter API.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>