ttbd: cleanup daemon sees inside power rails to be able to power down all
Our current definition of on or off was quite limited; if anything in
a power rail was off, the whole target was considered off and thus the
cleanup thread would skip it.

This would leave targets that were half up all the time up. A waste.

So introduce finer ways to get power at different granularity level,
power_rail_get(), which returns the state of each power component;
power_get() works the same as before, but the cleanup daemon looks at
anything powered up and powers it down when unused.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>