tcfl/zephyr: fix sanity check subcase detection
A change was made to scripts/sanitycheck that alters how it reports
subcases and flattens the report of subcases from the YAML to the same
level as those scanned from the source.

Change makes things way more complicated, so we need to add some ugly
heuristics to be able to tell which testcase things come from (the
YAML or the source code).

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>