tcfl: another approach to sanity check subtest case scanning
The previous approach we had still had holes in certain corner cases;
thus replaced with a version that seems to be working better.

Instead of trying to worl from the YAML list to the subcases, just go
over the list sanitycheck --list-tests and eliminate options starting
with whatever comes from the YAML file; consider the rest subcases.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>