tcfl.pos: fix rsync_np() and add deploy_path()
rsync_np needs to be passed a parameter so we can use it to transfer
paths or files indistictively (so it appends /. or not to paths).

When creating the destination paths, it has to prefix /mnt, otherwise
it tries to create it in the Provisioning OS space.

Also, fix the docs on how to pass parameters to deploy_*

Add deploy_path, which can rsync a file too, not just a tree; probably
we need to remove _tree.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>