tcf-image-setup: add image metadata to support more complicated cases
Allows to specify the filesystem to be created (instead of defaulting
to ext4) and post install scripts.

Gathers the rootfs type to put it in the metadata. Removes old
workarounds for manipulating the fstab when we had no post steps in
the metadata.

Fix little issues like always initializing $boot_part; few whitespace cleanups

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>