tcfl.pos: introduce and use image metadata
An image can provide a .tcf.metadata.yaml file in it's top level root
fs with indications about what file system format to use and post
rsync steps to further configure.

This is needed to be able to support more complex images and to remove
steps from tcf-image-setup.sh to setup serial consoles and such.

A schema file to verify the metadata yaml is introduced; upon
deploying the image, the metadata is loaded (if it exists).

Also _fsinfo_load() uses `lsblk` to get filesytem details to check if
a filesystem has to be reformatted

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>