tcfl/pos-uef: manually create local boot entry
Relying on bootctl to create the boot entry creates a window for a
race condition that might not allow us booting from the network to
control the system.

Thus replace with it installing the bootloader but we manually setting
up the EFI boot variables, entries and order only after everything is
in place and keeping network always first.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>