tcfl/pos_uefi: recycle the /boot partition
Before we were always blasting the /boot partition, and that takes
some time.

So now try to mount it, or fix it and rsync kernels in. Clean up older
kernels if the free space gets over a certain mark.

rsyncing means that when we are flashing something that is there
already, it'll take almost no time.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>