tcfl/clear-bbt: rework .t scanning to get early subcase list
Not having an early subcase list messes up accounting if the testcase
blocks or errors before it can be obtained.

Pre-parse the .t file using the same code as the bats tool itself to
gain an idea of the subcase list early on, following the same patter
as the Zephyr's sanity check subcases.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>