tcfl.pos: add -c to rsync
We were running rsync in non-c mode, which means it was only using the
timestamp and the size to determine if a file is the same or has
changed and needs updating.

Clearly not enough, as we found there were files with the same size
and mtime so they were not being updated.

-c is slower, but will be much more reliable

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>