tcfl.pos.deploy/rsync: fix arg passing issues and messages
Argument passing of a single string was broken; fixed.

Reduce verbosity level of messages so we see information about what's
happening in the deploy phase of all the paths being deployed.

Hardcode increase the timeout to rsync paths to the target; this is
not ideal, we need something better, but not sure what that'd be.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>