tcfl/report.py: Replace colon in report name with underscore
The colon (:) in report name, e.g. report-XXXX:YYYYYY.txt
is causing a lot of unnecessary problems. For example, it is
not possible to open it in gedit, it is being misinterpreted
as separators by some other programs. Sometimes program
treats colon as a special character with special meaning,
and steps such as escaping, putting it in quotes or
substitution with unicode representation is needed to
get inform the program to not treat it as special
character.

The use of colon is not really compulsory here, so substitute
it with underscore.

Signed-off-by: Ooi, Cinly <cinly.ooi@intel.com>