things: clean up and move to new interface schema
This yanks the old things interface that was embedded in class
ttbl.test_target and replaces it with the new schema where it is added
dynamically. It gets rid of a lot of duplicated code.

No backwards compat it maintained, since it is currently unused.

Client side changes to move everything to tcfl/target_ext_things.py
and registering the exception. Lots of ducplicated code are
consolidated from tcf and tcfl/ttb_client.py.

In the daemon side, API entry point in ttbd is removed in favour of
the ttbl.tt_interface method; all entry points in ttbl.test_target are
moved to ttbl.things.interface, cutting out a lot of duplication.

ttbl/{pc_ykush,usbrl08b,tt_qemu*} drivers ported, but not tested.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>