tcfl.pos: improve fsinfo loading
We were loading fsinfo in a blind way, which meant that sometimes it
didn't work and we went to either block or reinitialize.

Change it so we verify we got the information, waiting and retrying if
not; sometimes it seems it takes the kernel some time to get the info.

Add a generic API to get fsinfo data without having to know much
details of the structure.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>