ttbl.qemu: cleanup QEMU virtual target implementation
Refactor everything to remove the tt_qemu classes, so now QEMU is
an object that exposes power control interface to start/stop, exposing
debug and imaging interfaces. Much cleaner and reusing existing
infrastructure.

Refactor all the defaults configuration so QEMU VMs for POS and Zephyr
are created in a cleaner way.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>