tcf: cleanup the tunnel interface
Move the tunnel interface from the old hardcoded location in ttbd and
class test_target to a separate interface; add it by default upon
target insertion.

Add a healtcheck to the client and modify the client side to call the
new locations and to use the latest client API client conventions.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>