tcfl.target_ext_console: report console matches/mismatches
When a console match is found, the reported dictionary must indicate
the console output where the match was found.

Using the new generator_factory_c, report it so that the content can
be iterated vs. attached (which could put a lot of memory pressure).

As well, integrate into the new timeout handling mechanism for the
expectation, so that the message is more specific to a text timeout.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>