ttbl.allocation: release hooks are called upon allocation timeout
We were missing that when the allocation is timed out in the
ttbl.allocation.allocdb_c.delete() path, it is safest to just cleanup
and wipe the targets when they were active.

There was also a typo in calling state_get() to compare to an active
reservation.

Downgraded some FIXME messages so we don't cause false negatives on
checking server output.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>