ttbd: use list()/dict() vs [] / {}
We have had a few instances of the []/{} instantiated statically and
then modified in a shared environment when it was thought it was
not.

Although it is meant to be static to the local environment and that is
clear, for clarity and to avoid false confusions, just rewrite them
all to signify a new instance is being created.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>