commonl: fix dict_to_flat() overriding of data
When adding an empty dictionary (empty_dict == True), do not append
the value itself, but a pure empty dictionary, since if it is updated
later it will modify the source. In this case, create an empty dict()
to add.

This was happening when for example starting with the tags and then
adding data from the FSDB, then the tags were being modified.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>