tunnels: multiple fixes and clean ups
Main fix is to deprecate the list() call, since the inventory at
interfaces.tunnel contains the information needed for reporting the
current tunneling state. This also complies with the API spec.

Restructure list() to return those dictionaries instead of the hacked
up list of tuples it did now

On the way there:

- cleanup the messages in healthcheck

- add a couple of unit tests

- fix existing unit tests (use 127.0.0.1 instead of the FQDN, which
  doesn't really work)

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>