tcfl.biosl: fix failure to go to boot manager menu
In some cases the state machine got stuck when trying to go to an
specific menu because it's recovery mechanism, trying to redrwraw by
going up and down, was getting the state machine confused.

So let it just go up and retry when timing out, this is much better.

Note we also wrap all the entry select into a new function,
entry_select() that will be expanded later -- for now it adds some
pause to the process.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>