ttbl.allocation: fix but which was causing allocations to hang
The queue identifier for the target was using not enough data to
identify the queue entry and if multiple entries were submitted in the
same second, the timestamp used to identify them was being overriden
and thus the previous requests submited that second were dropped.

Add the allocation ID to add enough entropy so that it can never be
overriden.

Document bits and pieces around the area.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>