tcfl.tc: add tcfl.rc.subcase() decorator
This decorator allows to make a method a subcase, so that it is
reported as a subcase of the top level testcase.

The tcfl.tc.subcase() subcase name will by default be called after the
method name (without a phase prefix), otherwise a name can be given.

It relies on support provided by the msgid_c() context handler
(subcase_base support) and the break_on_non_pass property to allow
subcases to not cause the execution to halt completely on error.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>