tcfl.report_driver_c.report: fix argument list and data reporting
The argument list to report was packing reporter to be the testcase or
the target, which made it pretty confusing for the driver to tell,
especially when reporting in subcases and targets.

So change the API to pass the testcase (or subcase) and if reporting
from a target, the target reporting on.

This also enables the drivers to, when reporting DATA tags, add target
information in the proper space so we don't lose the target that
generated the data.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>