target.console: remove max_size due to futility and possible bugs
The max_size limit existed to try to limit how much data we were
chewing from the input, and it was a relic of a previous time.

However, it was impractical since at the time we were applying, we
already had read the remote data, so the local time ot process would
be negible in comparison. As well, it introduced bugs in offset
handling when it hit.

The likelyhood if it hitting was low since the limit was set by
default to 65k, but in any case -- not used, not needed, wipe.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>