ttbl.quartus: workaround/fix jtagd lingering process bug
sometimes the jtagd process seems to be changing names and when we try
to kill it at that point, we miss it.

It stays lingering, and so that the daemon_c.on() implementation can
kill it, we need to set also the path, so it tries to kill it with
that name.