tcfl.tc.expectation_c.on_found: fix exceptions not being raised when ERROR/command not found
Expectations that were set to catch exceptional situations (shell
ERRORs, command not found, etc) where not raising exceptions.

This was caused because the evaluation of the exception to bool was
being negative. This is because some exception types can bool to False
for whichever reason.

If there is nothing to raise, it's because there is a None, so compare
against that.