ttbd/ttbd get_json behaviour change
in endpoints with GET we give the user the option to send a json body, however
the way we were checking if the user has sent one, was buy just calling
the get_json function and expect it to return None.

This behaviour changed in newer versions of werkzeug
https://github.com/pallets/werkzeug/commit/bf4ede275335644102cd72c421c2b6976bc69883

Now get_json, expects that the content-type of the request to be
application/json. We no longer can just expect the function to
return None. If called with diff content-type it will raise an
exception.

I think the best way to fix this is by just telling get_json to fail in
silence. This way if the headers are not what he likes, it will just
return None (the expected old behaviour) then we can continue to see if
the user sent a form and continue with the current flow.

We only need to change this in GET endpoints, since the other ones are
expected to have a body.

Signed-off-by: benjamin santana <benjamin.santana.velazquez@intel.com>