tcfl.pos: boot_to_pos receives the boot interconnect object
boot interconnect is the network we use for booting, as specified in
the inventory key pos.boot_interconnect. It might not be defined, it
might be defined.

The boot_to_pos() functions before where just hacking access to the
boot interconnect information when need it. This change passes the
target object that represents the boot interconnect to undo those
hacks.

This cleans up some of the linux kernel commandline for PXE booting
keyword expansion.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>