tcfl.pos: move to use POS KWS in the boot_to_pos path
When calling all the boot to POS functions, first load all the POS
related keywords using kws_vars_load(). This allows for a consitent
handling of all those variables (getting them from the new location in
the inventory tree, maintaining backwards compat and allowing override
from the environment).

This is quite a large commit which boils down to:

- add kws_pos argumentt o the boot-to-pos path (*boot_to_pos*)

- remove from those functions all the special POS variable loading and
  handling

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>