tcfl.ui_cli_power: add new simplified power management CLI implementation
This will replace the implementation in tcfl/target_ext_power; uses
the new target discovery API and patterns in tcfl.ui_cli to implement
the command's functionalities.

It is way faster to get going and can run on any number of targets in
parallel.

Will cut a lot of duplicated code once the old version is folded.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>