ttbd: add timed/static allocations
This commit adds a field to the allocation called *endtime*.

When this field is *None* (default), the allocation behave as of now:
if they are idle, they are released by the daemon after a configured
amount of time.

However when this field is the string *static*, the allocation is
maintained until a user with enough privilege remove it.

If the field is a UTC timestamp string (YYYYmmddHHMMSS), the
allocation is deleted when that time arrives (or before if a user with
enough privilege removes it).