tcf/ui_cli: introduce processes_guess
Unify the handling of "how many subprocesses" to use in the servers
code with a utility to guess how many we shall use. This is fed into
run_fn_on_each_server() so it can take care of the guessing for the
caller (who can also override it if they know).

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>