tcfl.targets: fix issue when summarizing
When summarizing (simpliying SERVER/TARGETNAME as TARGETNAME), which
we do when there is only a single target called TARGETNAME, we were
hitting a bug where some entries where not summarized.

It turns our sometimes we were having multiple entries for
SERVER/TARGETNAME in self.rts_fullid_sorted which were not being
caught.

So ensure that doesn't happen.