ttbl.ui: refactor obtaining values from interconnects, support no networks
Targets that support multiple networks will have multiple IPs, MACs,
etc. Streamline how to show them, eg:

TARGETNAME    IPv4   MAC

or

TARGETNAME    nw1:IPv4.1     nw1:MAC.1
              nw2:IPv4.2     nw2:MAC.1
              nw3:IPv4.3     nw3:MAC.1

or

TARGETNAME    n/a            n/a

if it connects to no networks

This is also preparing for when we can, as user, configure the fields
we want in the inventory list.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>