ttbd.ui: add mechanism for offloading long fields to a tooltip
The python code can check fields that have the potential to be too
long with function short_field_maybe_add(), which adds a
FIELDNAME_short version.

The template code then can choose how to display the short vs the long.

Template for allocations changed for this, next commit will add the
call to short_field_maybe_add() as part of a refactor.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>