ui/allocations: added css and formatted html
did not add almost anything to the html, just did a refactoring for
formatting/style purposes, to make it a bit more readable, keep in mind
we are still talking about html so it can not be "super readble", but I
think this makes it a bit more.

on content I just splitted different web components into subsections
group, for them to be in different rows.

feedback is welcomed

Signed-off-by: bsantana <benjamin.santana.velazquez@intel.com>