ttbl.certs: initialize cert subsys on demand
Until now, we always initialized certs, which meant for every
allocation, we created a CA yes or yes. This has a cost and in most
cases, was going to waste since we were not using them--eg, everytime
local acquires to verify if the target needs to be powered off.

This moves the cert system to be on demand; if we create a certificate
(by means of using the PUT/certificates call), this means we need a
cert --then we bring up.