ttbd/conf_00_pos: added import & doc issue
while following the instructions I came across a example with an error.
If you declare a network name longer than 3 chars the daemon will not
start and will complain smth like:

AssertionError: nw_name must be one or two letters in the set [a-zA-Z], got
'nwa'

Fixed the doc example to just be able to copy and paste.

Also was missing ttbl.rsync at the imports, added it

Signed-off-by: Santana, Benjamin <benjamin.santana.velazquez@intel.com>