ttbl/ui: adding lazy loading
modified how we render power rail, to just get the components
without their state from the inventory, this also means some changes in
target.html (remove the state from the initial rendering of the page).

main.js: we also added a function to do a /power/list call to the
server. This function will update the table cell that tells us the power
state of each component. More details on the impl in the docs

Signed-off-by: benjamin santana <benjamin.santana.velazquez@intel.com>