ttbd/ui/targethtml: not acquire no display buttons
if the target is not acquired we do not want to display the actions you
could do. What we did was basically recreate the same control panel but
with all the buttons disabled. This ended up creating unnecessary
redundancy making it super easy to be out of sync. Instead of that we
just display a message telling the user to acquire the machine.

Signed-off-by: benjamin santana <benjamin.santana.velazquez@intel.com>