ttbd/ui: added more info on the consoles
we did not have an easy way from the ui to tell if a console was
disabled or enabled. This fix adds beside the console info its state.

It also dims the button of the current state of the console. Meaning if
the console is enabled the enable button will be dimmed.

Signed-off-by: benjamin santana <benjamin.santana.velazquez@intel.com>