tcf/tc: move usage of tcfl.ttb_client to tcfl.targets
This removes the last user of the old tcfl.ttb_client core discovery
code to the new one based on tcfl.targets; some parts are a bit of a
hack (especially _target_discovery) but that code is also going to be
improved later on.

There are still a few users of the API in self.rtb, so it is kept for
the time being as future small commits wipe them.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>