tcf: remove deprecated tcfl.ttb_client
This was the original low level core module to access the server,
which has served good but it has reached limitations.

It was deprecated long ago by the new tcfl.server_c class, and the
tcfl.servers module; code has been slowly migrated to use it and the
final bits have been moved in previous commits.

This final commit removes the leftover imports and aesthetic users and
the code itself.