ui/static/js/targets.js: new js file 4 targets.js
the user will be able to acquire systems from targets.html but wont be
able to do much else, there is no sense in loading EVERY function in the
main.js file.

It is cleaner to just load a new file, targets.js, with just the
function needed. And adding that file to the headers of both
targets.html and target.html.

Also created a mini css file for targets.html, with the same reasoning.
With time we would have the css files in a modular fashion. Just loading
the selectors each html file needs. This is the first step into that
direction.

Signed-off-by: benjamin santana <benjamin.santana.velazquez@intel.com>