TCF: add initial Jinja2 reporting template in HTML
This adds the ability to generate HTML reports of execution of test
cases:

- report-*html: Jinja2 templates for rendering of execution info
- setup.py: ensure templates are installed
- report_jinja2: configure templates into reporting engine so reports
  are made in HTML

Note this is still very crude; uses basic 2000-era HTML construct and
there is a long list of ways it can be improved. This is just meant as
a proof-of-concept and as such is disabled by default.

To enable, add to a TCF configuration file:

>>> tcfl.report_jinja2.driver.templates['html']['report_fail'] = True
>>> tcfl.report_jinja2.driver.templates['html']['report_block'] = True
>>> tcfl.report_jinja2.driver.templates['html']['report_error'] = True