WebUI: add servers_info structure to generate list of known servers
added new list item to navbar that displays the list of all available
servers. We store them in cache 20 min to not have the operation of
discovering the servers occur everytime one loads a webpage.

Added style in css for this dropdown/list.

Most of the work in this commit was done by inaky-intc, while I
(bsantana-intc) did the styling and minor improvements.

Signed-off-by: benjamin santana <benjamin.santana.velazquez@intel.com>