ttbd: use pure flat dictionary to feed to properies_set()
In the past, we were converting the properties passed in the endpoint
to a flat dictionary; however, by default dict_to_flat() returns a
mixed dict/flat dictionary due to historical reasons.

So force a pure flat dictionary so that we can feed deep data
structures properly converted to flat first, as expected by
properties_set().