commonl.buildah_image_create(): add ability to map extra directories
We'll use this to map DNF caches so when it takes too long to download
and the process timestout, a retry will just pick it up from where it
timedout before.

This happens first time we build the image: ttbd service starts, takes
too long to download it all, systemd kills ttbd, restarts and keeps
rebuilding without having to download from the get go.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>