ttbd: ensure the execution of the setup defer list sends keepalives
Some instances are timing out to start because the systemd keepalives
are never sent.

Before the cleanup thread starts sending systemd keepalives, we had to
execute the setup defer action list, that in some configs is used to
power on remote network infrastructure for deployment resilience.

However, if some of that was taking a long time because other issues,
this instance would timeout because systemd would think it was dead
and be restarted.

So to avoid this be a problem, allow the defer list executor to
keepalive every ten seconds while waiting for its tasks to finish.