ttbl.power.daemon_c: resolve executables
In Fedora 39, /usr/bin/socat is symlinked to /usr/bin/socat1, which
means that when we try to detect it, we fail.

This means that then the server will try all kinds of things to
recover, restart, etc and always fails, which ends up beig a mess.

So we resolve the first component -- the executable -- so we don't
fail to detect it.