ttbl.power: another round of fixes for symlinked binaries
The previous half-witted attempt at solving the problem of symlinked
binaries was not enough; try to plug the niche cases where it was
failing (non absolute paths, pass through of binaries not
found). Acknowledge this is not the right fix but will do until we
have time to implement it.