ttbl.device_resolver: replace device->device_name, add device_[real]path
The *device* field is used in sysfs, so if we use it to host the
device name, we can't match against the sysfs one, so rename that to
device_name.

Add device_path and device_realpath, to be able to match against the
bus view and physical view of the device.