tcf/console/read: support chunk encoding, no Content-Length
When we enabled the gunicorn server, much better escalability and
performance than Tornado, the consoles stopped working correctly.

In our protocol, we relied in Content-Length to determine the size of
the read -- but gunicorn uses HTTP/1.1 with Type-encoding: chunked,
which disallows using Content-Length. So switch to be able to
recognize the length size based on what the HTTP client library
reconstructs.

Update docs.