tcfl.targets.discover_agent_c: shorten names in rt entries too
We shorten names (make SERVERAKA/TARGETID just be TARGETID) when
TARGETID is unique (eg: avaialble in a single server.

However, we were doing that just in the top level indexes (tcfl.rts
and tcfl.rts_flat) and those specific to a discovery agent. We were
missing to shorten the *fullid* field in the dictonary with each
target's data.

This was causing things like "tcf run" to always use the long name,
adding unnecessary output.