tcfl.tc._targets_discover: fix tcf run not finding targets when no verbosity
When calling tcfl.targets.setup_by_spec(), it uses the verbosity
argument as a hint to decide how many fields to query from the
inventory (to speed up when only a few are needed).

When calling tcf run with no -v, we were passing that up and getting
only the names from the inventory, which meant we couldn't match on
the conditions.

*tcf run* needs ALL the fields, since we don't know what scripts will
do; so just call it with verbosity 1 to ensure all fields are
obtained.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>