ttbd/ui/static/css/button: new button design
this new design not only looks better (kind of subjective) but it
introduces a new structure for the css file. Instead of having a bunch
of classes for different buttons, we create one standard for the whole
page, and create small variations to specific things.

- no class: if no class defined, a simple white button with hover
  and active animations will be displayed. Perfect for every "normal"
  button in the page.

- primary: this variation is for buttons that need to be highlighted,
  and users to be drawn to them. It is displayed in blue.

- secundary: this variation is for buttons that do not need to draw the
  user, but require a bit more of attention by the user if they are
  going to press them, this is highlighted in light blue

- warning: red big button. This is for buttons that might delete
  something, or perform operations that user must be careful

that's it, 4 diff types of buttons for all the web page. Of course we
can add more if needed, but this will be sufficient for the most part.

Signed-off-by: benjamin santana <benjamin.santana.velazquez@intel.com>