ttbd/setup.py: add path to /share/tcf/ui/
In the webui, we are using some external libraries (css and js files).

Before 7c0288c1 we had them mixed with our static files, making it a bit
confusing for the devs which file we did ourselves and which file is
from an external library.

Of course, all the external libraries are credited under
ttbd/ui/source.yaml but it is faster to identify them if we have them
in diff dirs

We moved all the external files to `ttbd/ui/static/js/external` and
`ttbd/ui/static/css/external`. Which leads to this commit.

When doing the installation process we were using the glob
`glob.glob('ui/static/css/*')` to add all files under css, but that glob
was having issues picking up the dirs inside the top level path.

Therefore we just add entries for those specific subdirs like:
`glob.glob('ui/static/css/external/*.css')`

Signed-off-by: Santana, Benjamin <benjamin.santana.velazquez@intel.com>