ttbd/UI: refresh server discovery periodically so UI loads faster
In the UI, we need the other server's information so the user can
easily switch. We load it synchronously, so it can take time if it has
to refresh.

Instead of making the UI code async and more complex, we have the
server refreshed frequently from the cleanup thread and thus the user
will have the info always in there.