tcf/alloc-ls: allow listing only given ALLOCIDs
The current argument for this command was a targetspec which was
inneficient, since it ended up listing always all allocations.

Move the targetsec to -t|--target; enable tcfl.allocations.ls() to
take a list of allocids to filter for. Rework the command line
arguments for the main arguments to be allocation IDs.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>