commonl.dict_to_flat(): fix bug which resulted in missing fields
The implementation was missing fields when they were deeper in a deep
dictionary, so added extra checks to detect that condition and include
them.

This started because something as simple as:

  $ tcf get TARGETNAME -p instrumentation

was still printing a lot of the fields from the tags--turns out that
this function wasn't filtering good enough.

Added a test case which can have plenty of test vectors easily added,
in case we find more situations that were not covered properly.