ttbl.raritan_emx: allow setting PDU URL/outlet/password in database
Currently, the Raritan EMX driver to power control the PDU would take
what PDU and outlet to use from config info, we could not update it
realtime.

This patch allows that info to come from config (current behaviour)
and also from the inventory DB (from fields
instrumentation.CODE.{url,password}).

It also allows to encode most info in the URL in the forms

        [https://][USERNAME[:PASSWORD]@]HOSTNAME[:OUTLETNUMBER]

note if PASSWORD spec has /, it has to be specified separate; this is
why there is a separate password field in the inventory DB.

The move implied creating a separata function _url_resolve() that takes
the URL and the password, defaults to config if None and then
parses/resolve passords, figutes the ourlet number and then just
returns the parsed URL, the password to use and the outlet number
which the caller can use directly.

We use that in all the key functions that use the API (on/off/get) to
create the RPC API object as needed, since we don't know when that
data was updated we can't keep the object around.

To call the power consumption agent, we just pass the parsed data to
the agent's execution.

Note the bulk of the commit is the testcases to verify the
implementation as much as possible (taking into account we do it
without a PDU actually available).