tcfl.server_c: mark places of failure with _record_failure()
This pretty much is in herds_get() and _targets_get(); whenever we
detect a failure, we record it.

We also update its implementation, adding microsecond resolution
(since seconds is not enough to detect failures in the same second);
also, clear failures if we record a success.

Follow up commit will act on the recorded data.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>