UI: ttbl: fix allocation removal on guest removal
When a user pressed "remove me as a guest" in the UI, it was causing
allocations to be removed when the user was also an admin.

The problem is we were calling "allocation removal" when removing as a
guest, that will remove a user that is a guest; but a user that has
admin permissions will just kill the allocation.

So change the "remove me as a guest" to call pure guest removal
instead. Update doc tool tips about it too.

Note we also had to clean up the way we call JS common_error_check(),
since we were not able to get response text for the error -- and since
it needs to be obtained from a promise, it can only be invoked in
functions that are async.