ttbd/UI: fix "add myself as a guest"
There was a race condition since the request was being made and
inmediately we were trying to reload. The request was being killed and
the action not taken.

So change to make the request sync and only once it completes, we
reload the window to see ourselves as a guest that can use.