ttbl.images: keep accessing files while flashing from them
When we have flashing processes that are very long, the flashing tool
might not access it while it flashes but then need it later and try to
access again (eg: for a verification pass).

However, if the file we are flashing is in a dynamically mounted
directory, the underlying layer might see the file hasn't been
accessed and start unmounting things under the driver, resulting in a
non-accessible file.

While It Should Not Happen(tm), we have no control over how the tools
do things, so this implementation basically just stats the files while
we are faiting for the flashing process to complete, so that the file
system layer sees some activity and maybe not trigger a dynamic file
system umount.