tcf/guest-*: allow adding users by targetid instead of allocationid
Normally when we want to guest someone from the command line, we need
to:

  $ tcf ls -v SUTNAME
  SUTNAME [USERNAME:ALLOCID]

copy the ALLOCID and run

  $ tcf guest-add ALLOCID [USERNAME]

which is a pain after a while

This introduces using -t to specify a targetspec that will find the
allocation ids of the targets that match the spec and guest the user
in there (or remove or list):

  $ tcf guest-add -t TARGETSPEC [USERNAME]
  $ tcf guest-rm -t TARGETSPEC USERNAME
  $ tcf guest-ls -t TARGETSPEC

If TARGETSPEC matches multiple targets, it'll get the allocations for
each and do the operation in them.

So for example, to add myself to whatever active allocation is in
target q11j:

 $ target guest-add -t q11j

likewise to remove myself;

 $ target guest-add -t q11 self

The implementation basically lists all the targets and then collects
all the allocations on each server, then send the guest/add/rm to each
server and allocation.